# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_root","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/offers":{"get":{"operationId":"OffersController_listOffers","parameters":[{"name":"lat","required":true,"in":"query","schema":{"type":"number","example":41.3275}},{"name":"lng","required":true,"in":"query","schema":{"type":"number","example":19.8187}},{"name":"radius_km","required":true,"in":"query","schema":{"type":"number","example":5}},{"name":"from","required":true,"in":"query","schema":{"type":"string","format":"date-time","example":"2026-01-10T08:00:00Z"}},{"name":"to","required":true,"in":"query","schema":{"type":"string","format":"date-time","example":"2026-01-10T20:00:00Z"}}],"responses":{"200":{"description":"List offers (geo bounding-box + time window)","content":{"application/json":{"schema":{"type":"object","properties":{"items":{"type":"array","items":{"type":"object"}},"meta":{"type":"object","properties":{"lat":{"type":"number"},"lng":{"type":"number"},"radiusKm":{"type":"number"},"from":{"type":"string","format":"date-time"},"to":{"type":"string","format":"date-time"},"count":{"type":"integer"}}}}}}}}},"tags":["Offers"]}},"/vendor/customers/blocked":{"get":{"operationId":"OrdersController_blockedCustomers","parameters":[],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/orders":{"post":{"operationId":"OrdersController_reserve","parameters":[],"responses":{"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}/cancel":{"patch":{"operationId":"OrdersController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/vendor/orders/{id}/claim":{"post":{"operationId":"OrdersController_claim","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Orders"]}},"/vendor/orders":{"get":{"operationId":"OrdersController_vendorOrders","parameters":[{"name":"vendorId","required":true,"in":"query","schema":{"type":"string"}},{"name":"status","required":true,"in":"query","schema":{"type":"string"}},{"name":"take","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/payments/mock/succeed":{"post":{"operationId":"PaymentsController_succeed","parameters":[],"responses":{"201":{"description":""}},"tags":["Payments"]}},"/me/self-unblock/bank-transfer/init":{"post":{"operationId":"SelfUnblockController_init","parameters":[],"responses":{"201":{"description":""}},"tags":["SelfUnblock"]}},"/me/self-unblock/bank-transfer/{paymentId}/proof":{"post":{"operationId":"SelfUnblockController_proof","parameters":[{"name":"paymentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["SelfUnblock"]}},"/admin/self-unblock/{paymentId}/confirm":{"post":{"operationId":"SelfUnblockController_confirm","parameters":[{"name":"paymentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["SelfUnblock"]}},"/admin/expirer/tick":{"post":{"operationId":"ExpirerAdminController_tick","parameters":[],"responses":{"201":{"description":""}},"tags":["ExpirerAdmin"]}},"/admin/users/role":{"post":{"operationId":"AdminUsersController_setRole","parameters":[],"responses":{"201":{"description":""}},"tags":["AdminUsers"]}},"/vendor/me":{"get":{"operationId":"VendorsController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Vendors"]}},"/vendor/me/{vendorId}":{"patch":{"operationId":"VendorsController_updateMyVendor","parameters":[{"name":"vendorId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vendors"]}},"/vendor/apply":{"post":{"operationId":"VendorsController_apply","parameters":[],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/vendor/offers":{"get":{"operationId":"VendorsController_listMyOffers","parameters":[],"responses":{"200":{"description":""}},"tags":["Vendors"]},"post":{"operationId":"VendorsController_createOffer","parameters":[],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/vendor/offers/{id}":{"patch":{"operationId":"VendorsController_updateOffer","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vendors"]}},"/vendor/offers/{id}/publish":{"post":{"operationId":"VendorsController_publish","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/vendor/offers/{id}/pause":{"post":{"operationId":"VendorsController_pause","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/vendor/offers/{id}/resume":{"post":{"operationId":"VendorsController_resume","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/admin/vendors/status":{"post":{"operationId":"VendorsController_setVendorStatus","parameters":[],"responses":{"201":{"description":""}},"tags":["Vendors"]}},"/admin/users":{"get":{"operationId":"AdminOpsController_users","parameters":[{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["AdminOps"]}},"/admin/vendors":{"get":{"operationId":"AdminOpsController_vendors","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}},{"name":"ownerUserId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["AdminOps"]}},"/admin/offers":{"get":{"operationId":"AdminOpsController_offers","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}},{"name":"vendorId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["AdminOps"]}},"/admin/offers/status":{"post":{"operationId":"AdminOpsController_setOfferStatus","parameters":[],"responses":{"201":{"description":""}},"tags":["AdminOps"]}},"/admin/otp/stats":{"get":{"operationId":"OtpAdminController_stats","parameters":[],"responses":{"200":{"description":""}},"tags":["OtpAdmin"]}},"/admin/otp/cleanup":{"post":{"operationId":"OtpAdminController_cleanup","parameters":[],"responses":{"201":{"description":""}},"tags":["OtpAdmin"]}},"/admin/audit":{"get":{"operationId":"AdminAuditController_list","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"actorUserId","required":true,"in":"query","schema":{"type":"string"}},{"name":"action","required":true,"in":"query","schema":{"type":"string"}},{"name":"targetType","required":true,"in":"query","schema":{"type":"string"}},{"name":"targetId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["AdminAudit"]}},"/admin/dev/impersonate":{"post":{"operationId":"DevAuthController_impersonate","parameters":[],"responses":{"201":{"description":""}},"tags":["DevAuth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/request-otp":{"post":{"operationId":"AuthController_requestOtp","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}}},"info":{"title":"Savora API","description":"Savora backend (api_nest)","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{}}}